# ContentApiCore/wwwroot/json/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Open Data Hub Content Api",
    "description": "Open Data Hub Content Api based on .Net Core with PostgreSQL",
    "termsOfService": "https://opendatahub.readthedocs.io/en/latest/",
    "contact": {
      "name": "Open Data Hub Team",
      "url": "https://opendatahub.com/",
      "email": "help@opendatahub.com"
    },
    "version": "v1"
  },
  "servers": [
    {
      "url": "https://api.tourism.testingmachine.eu"
    }
  ],
  "paths": {
...
    "/v1/Example": {
      "get": {
        "tags": [
          "Example"
        ],
        "summary": "GET Example List",
        "parameters": [
          {
            "name": "language",
            "in": "query",
            "description": "Language field selector, displays data and fields in the selected language (default:'null' all languages are displayed)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pagenumber",
            "in": "query",
            "description": "Pagenumber",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pagesize",
            "in": "query",
            "description": "Elements per Page, (default:10)",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "exampletype",
            "in": "query",
            "description": "Example Types Filter (Separator ',' List of ExampleTypes IDs), (default:'null')",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "idlist",
            "in": "query",
            "description": "IDFilter (Separator ',' List of IDs), (default:'null')",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "langfilter",
            "in": "query",
            "description": "Language filter (returns only data available in the selected Language, Separator ',' possible values: 'de,it,en,nl,sc,pl,fr,ru', 'null': Filter disabled)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "active",
            "in": "query",
            "description": "Active Filter (possible Values: 'true' only Active Data, 'false' only Disabled Data), (default:'null')",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "updatefrom",
            "in": "query",
            "description": "Returns data changed after this date Format (yyyy-MM-dd), (default: 'null')",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "seed",
            "in": "query",
            "description": "Seed '1 - 10' for Random Sorting, '0' generates a Random Seed, 'null' disables Random Sorting, (default:null)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "publishedon",
            "in": "query",
            "description": "Published On Filter (Separator ',' List of publisher IDs), (default:'null')",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "source",
            "in": "query",
            "description": "Filter by Source (Separator ','), (Sources available 'idm','noi'...),(default: 'null')",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fields",
            "in": "query",
            "description": "Select fields to display, More fields are indicated by separator ',' example fields=Id,Active,Shortname (default:'null' all fields are displayed). <a href=\"https://github.com/noi-techpark/odh-docs/wiki/Common-parameters%2C-fields%2C-language%2C-searchfilter%2C-removenullvalues%2C-updatefrom#fields\" target=\"_blank\">Wiki fields</a>",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "searchfilter",
            "in": "query",
            "description": "String to search for, Title in all languages are searched, (default: null) <a href=\"https://github.com/noi-techpark/odh-docs/wiki/Common-parameters%2C-fields%2C-language%2C-searchfilter%2C-removenullvalues%2C-updatefrom#searchfilter\" target=\"_blank\">Wiki searchfilter</a>",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "rawfilter",
            "in": "query",
            "description": "<a href=\"https://github.com/noi-techpark/odh-docs/wiki/Using-rawfilter-and-rawsort-on-the-Tourism-Api#rawfilter\" target=\"_blank\">Wiki rawfilter</a>",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "rawsort",
            "in": "query",
            "description": "<a href=\"https://github.com/noi-techpark/odh-docs/wiki/Using-rawfilter-and-rawsort-on-the-Tourism-Api#rawsort\" target=\"_blank\">Wiki rawsort</a>",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "removenullvalues",
            "in": "query",
            "description": "Remove all Null values from json output. Useful for reducing json size. By default set to false. Documentation on <a href=\"https://github.com/noi-techpark/odh-docs/wiki/Common-parameters,-fields,-language,-searchfilter,-removenullvalues,-updatefrom#removenullvalues\" target=\"_blank\">Opendatahub Wiki</a>",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ExampleLinkedJsonResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExampleLinkedJsonResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExampleLinkedJsonResult"
                }
              },
              "text/csv": {
                "schema": {
                  "$ref": "#/components/schemas/ExampleLinkedJsonResult"
                }
              },
              "application/ld+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExampleLinkedJsonResult"
                }
              },
              "application/ldjson": {
                "schema": {
                  "$ref": "#/components/schemas/ExampleLinkedJsonResult"
                }
              },
              "application/rawdata": {
                "schema": {
                  "$ref": "#/components/schemas/ExampleLinkedJsonResult"
                }
              }
            }
          },
          "400": {
            "description": "Request Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/csv": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/ld+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/ldjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/rawdata": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "oauth2": [ ]
          }
        ]
      },
      "post": {
        "tags": [
          "Example"
        ],
        "summary": "POST Insert new Example",
        "requestBody": {
          "description": "Example Object",
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/ExampleLinked"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExampleLinked"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ExampleLinked"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ExampleLinked"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PGCRUDResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PGCRUDResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PGCRUDResult"
                }
              },
              "text/csv": {
                "schema": {
                  "$ref": "#/components/schemas/PGCRUDResult"
                }
              },
              "application/ld+json": {
                "schema": {
                  "$ref": "#/components/schemas/PGCRUDResult"
                }
              },
              "application/ldjson": {
                "schema": {
                  "$ref": "#/components/schemas/PGCRUDResult"
                }
              },
              "application/rawdata": {
                "schema": {
                  "$ref": "#/components/schemas/PGCRUDResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/csv": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/ld+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/ldjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/rawdata": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Server Error"
          }
        },
        "security": [
          {
            "oauth2": [ ]
          }
        ]
      }
    },
...
    "/v1/JsonLD/DetailInLD": {
      "get": {
        "tags": [
          "JsonLD"
        ],
        "summary": "GET Detail Data in JSON LD Format (Schema.org Datatypes as output)",
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "description": "Data Type to transform currently available: ('accommodation', 'gastronomy', 'event', 'recipe', 'poi', 'region', 'tv', 'municipality', 'district', 'skiarea') required",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Id",
            "in": "query",
            "description": "ID of the data to transform, required",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "language",
            "in": "query",
            "description": "Output Language, standard EN",
            "schema": {
              "type": "string",
              "default": "en"
            }
          },
          {
            "name": "idtoshow",
            "in": "query",
            "description": "ID to show on Json LD @id, not provided Id of ODH api call is taken",
            "schema": {
              "type": "string",
              "default": ""
            }
          },
          {
            "name": "urltoshow",
            "in": "query",
            "description": "url to show on Json LD @id, not provided idtoshow is taken, idtoshow not provided url is filled with url of the data",
            "schema": {
              "type": "string",
              "default": ""
            }
          },
          {
            "name": "imageurltoshow",
            "in": "query",
            "description": "image url to show on Json LD @image, not provided image url of data is taken",
            "schema": {
              "type": "string",
              "default": ""
            }
          },
          {
            "name": "showid",
            "in": "query",
            "description": "Show the @id property in Json LD default value true",
            "schema": {
              "type": "boolean",
              "default": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "oauth2": [ ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Detail": {
        "type": "object",
        "properties": {
          "Header": {
            "type": "string",
            "nullable": true
          },
          "SubHeader": {
            "type": "string",
            "nullable": true
          },
          "IntroText": {
            "type": "string",
            "nullable": true
          },
          "BaseText": {
            "type": "string",
            "nullable": true
          },
          "Title": {
            "type": "string",
            "nullable": true
          },
          "AdditionalText": {
            "type": "string",
            "nullable": true
          },
          "MetaTitle": {
            "type": "string",
            "nullable": true
          },
          "MetaDesc": {
            "type": "string",
            "nullable": true
          },
          "GetThereText": {
            "type": "string",
            "nullable": true
          },
          "Language": {
            "type": "string",
            "nullable": true
          },
          "Keywords": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "ParkingInfo": {
            "type": "string",
            "nullable": true
          },
          "PublicTransportationInfo": {
            "type": "string",
            "nullable": true
          },
          "AuthorTip": {
            "type": "string",
            "nullable": true
          },
          "SafetyInfo": {
            "type": "string",
            "nullable": true
          },
          "EquipmentInfo": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ExampleLinked": {
        "type": "object",
        "properties": {
          "Self": {
            "type": "string",
            "description": "generated field",
            "nullable": true,
            "readOnly": true
          },
          "ODHTags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExampleTypes"
            },
            "description": "generated field",
            "nullable": true,
            "readOnly": true
          },
          "_Meta": {
            "$ref": "#/components/schemas/Metadata"
          },
          "Id": {
            "type": "string",
            "nullable": true
          },
          "FirstImport": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "LastChange": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "Source": {
            "type": "string",
            "nullable": true
          },
          "Active": {
            "type": "boolean"
          },
          "Shortname": {
            "type": "string",
            "nullable": true
          },
          "Mapping": {
            "type": "object",
            "additionalProperties": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            },
            "nullable": true
          },
          "HasLanguage": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "PublishedOn": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "Detail": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/Detail"
            },
            "nullable": true
          },
          "GpsInfo": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GpsInfo"
            },
            "nullable": true
          },
          "GpsPoints": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/GpsInfo"
            },
            "description": "generated field",
            "nullable": true,
            "readOnly": true,
            "deprecated": true
          },
          "ImageGallery": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ImageGallery"
            },
            "nullable": true
          },
          "ContactInfos": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/ContactInfos"
            },
            "nullable": true
          },
          "ExampleTypeIds": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "LicenseInfo": {
            "$ref": "#/components/schemas/LicenseInfo"
          },
          "AdditionalProperties": {
            "type": "object",
            "additionalProperties": { },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ExampleLinkedJsonResult": {
        "type": "object",
        "properties": {
          "TotalResults": {
            "type": "integer",
            "format": "int32"
          },
          "TotalPages": {
            "type": "integer",
            "format": "int32"
          },
          "CurrentPage": {
            "type": "integer",
            "format": "int32"
          },
          "PreviousPage": {
            "type": "string",
            "nullable": true
          },
          "NextPage": {
            "type": "string",
            "nullable": true
          },
          "Seed": {
            "type": "string",
            "nullable": true
          },
          "Items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExampleLinked"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ExampleTypes": {
        "type": "object",
        "properties": {
          "Id": {
            "type": "string",
            "nullable": true
          },
          "Self": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "GpsInfo": {
        "type": "object",
        "properties": {
          "Gpstype": {
            "enum": [
              "position",
              "viewpoint",
              "startingandarrivalpoint",
              "startingpoint",
              "arrivalpoint",
              "carparking",
              "halfwaypoint",
              "valleystationpoint",
              "middlestationpoint",
              "mountainstationpoint"
            ],
            "type": "string",
            "nullable": true
          },
          "Latitude": {
            "type": "number",
            "format": "double"
          },
          "Longitude": {
            "type": "number",
            "format": "double"
          },
          "Altitude": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "AltitudeUnitofMeasure": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ImageGallery": {
        "type": "object",
        "properties": {
          "ImageName": {
            "type": "string",
            "nullable": true
          },
          "ImageUrl": {
            "type": "string",
            "nullable": true
          },
          "Width": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "Height": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "ImageSource": {
            "type": "string",
            "nullable": true
          },
          "ImageTitle": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "nullable": true
          },
          "ImageDesc": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "nullable": true
          },
          "ImageAltText": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "nullable": true
          },
          "IsInGallery": {
            "type": "boolean",
            "nullable": true
          },
          "ListPosition": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "ValidFrom": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "ValidTo": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "CopyRight": {
            "type": "string",
            "nullable": true
          },
          "License": {
            "type": "string",
            "nullable": true
          },
          "LicenseHolder": {
            "type": "string",
            "nullable": true
          },
          "ImageTags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "LicenseInfo": {
        "type": "object",
        "properties": {
          "License": {
            "enum": [
              "CC0",
              "CC-BY",
              "Closed"
            ],
            "type": "string",
            "nullable": true
          },
          "LicenseHolder": {
            "type": "string",
            "nullable": true
          },
          "Author": {
            "type": "string",
            "nullable": true
          },
          "ClosedData": {
            "type": "boolean",
            "description": "readonly field",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "Metadata": {
        "type": "object",
        "properties": {
          "Id": {
            "type": "string",
            "nullable": true
          },
          "Type": {
            "enum": [
              "accommodation",
              "accommodationroom",
              "event",
              "odhactivitypoi",
              "measuringpoint",
              "webcam",
              "article",
              "venue",
              "eventshort",
              "experiencearea",
              "region",
              "metaregion",
              "tourismassociation",
              "municipality",
              "district",
              "area",
              "wineaward",
              "skiarea",
              "skiregion",
              "odhtag",
              "publisher",
              "tag",
              "weatherhistory",
              "weather",
              "weatherdistrict",
              "weatherforecast",
              "weatherrealtime",
              "snowreport",
              "odhmetadata",
              "package",
              "ltsactivity",
              "ltspoi",
              "ltsgastronomy"
            ],
            "type": "string",
            "nullable": true
          },
          "LastUpdate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "Source": {
            "type": "string",
            "nullable": true
          },
          "Reduced": {
            "type": "boolean"
          },
          "UpdateInfo": {
            "$ref": "#/components/schemas/UpdateInfo"
          }
        },
        "additionalProperties": false
      },
      "NotifierResponse": {
        "type": "object",
        "properties": {
          "Response": {
            "nullable": true
          },
          "HttpStatusCode": {
            "$ref": "#/components/schemas/HttpStatusCode"
          },
          "Service": {
            "type": "string",
            "nullable": true
          },
          "Success": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "PGCRUDResult": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "odhtype": {
            "type": "string",
            "nullable": true
          },
          "operation": {
            "type": "string",
            "nullable": true
          },
          "updated": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "created": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "deleted": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "error": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "errorreason": {
            "type": "string",
            "nullable": true
          },
          "compareobject": {
            "type": "boolean",
            "nullable": true
          },
          "objectchanged": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "objectimagechanged": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "pushchannels": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "changes": {
            "nullable": true
          },
          "pushed": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/NotifierResponse"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "Type": {
            "type": "string",
            "nullable": true
          },
          "Title": {
            "type": "string",
            "nullable": true
          },
          "Status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "Detail": {
            "type": "string",
            "nullable": true
          },
          "Instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "UpdateInfo": {
        "type": "object",
        "properties": {
          "UpdatedBy": {
            "type": "string",
            "nullable": true
          },
          "UpdateSource": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "oauth2": {
        "type": "oauth2",
        "flows": {
          "password": {
            "tokenUrl": "https://auth.opendatahub.testingmachine.eu/auth/realms/noi/protocol/openid-connect/token",
            "scopes": { }
          },
          "clientCredentials": {
            "tokenUrl": "https://auth.opendatahub.testingmachine.eu/auth/realms/noi/protocol/openid-connect/token",
            "scopes": { }
          }
        }
      }
    }
  },
...
}